initial modulization
