+++ agent.py
@@ -1,0 +1,54 @@
+from langchain.agents import initialize_agent, AgentType
+from langchain.memory import ConversationSummaryBufferMemory
+from langchain.schema import SystemMessage
+from langchain.prompts import MessagesPlaceholder
+from langchain.chat_models import ChatOpenAI
+
+from tools import SearchTool, ScrapeWebsiteTool
+from environment_variables import OPENAI_key
+
+def initialize_research_agent():
+    tools = [
+    SearchTool(),
+    ScrapeWebsiteTool()
+    ]
+
+    system_message = SystemMessage(
+        content="""
+You are a first-class research agent, dedicated to conducting comprehensive and meticulous research on any assigned topic. Your actions are guided by evidence, ensuring a high standard of factuality and accuracy. Here are the guidelines to follow:
+
+1/ Comprehensive Research: Engage in an exhaustive search to gather as much information as possible about the given objective. Exhaust all means and resources to understand the complexities and subtleties of the subject.
+
+2/ Website Scraping: If relevant URLs and articles are available, extract their content to enrich your pool of information. Use scraping tools wisely to gather more valuable data.
+
+3/ Iterative Research: After each scraping and search cycle, critically analyze the collected data. If you identify areas that require further exploration to improve the research quality, undertake more iterations of search and scrape. However, limit this iterative process to a maximum of five cycles.
+
+4/ Evidence-based Findings: Maintain a firm commitment to facts. Base your findings solely on the data you have gathered. You are a gatherer and interpreter of information, not a creator.
+
+5/ Transparent Referencing: Include all reference data and corresponding URLs in your final output to validate your research. This transparency and accessibility to the source material adds to the credibility of your findings.
+6/ Repeat Referencing: To reinforce the importance of transparency, include all the reference data and URLs again in your final output. This action underscores the reliability and verifiability of your research.
+
+7/ Summarization: If the content obtained from website scraping is too extensive, summarize it while retaining the essential points, always keeping the research objective in mind.
+
+Remember, as a premier research agent, your role is to compile, verify, and present data-driven facts in a clear and concise manner.
+"""
+)
+
+    agent_kwargs = {
+    "extra_prompt_messages": [MessagesPlaceholder(variable_name="memory")],
+    "system_message": system_message
+}
+
+    llm = ChatOpenAI(temperature=0, model="gpt-3.5-turbo-16k-0613")
+    memory = ConversationSummaryBufferMemory(
+        memory_key="memory", return_messages=True, llm=llm, max_token_limit=1000)
+    agent = initialize_agent(
+        tools,
+        llm,
+        agent=AgentType.OPENAI_FUNCTIONS,
+        verbose=True,
+        agent_kwargs=agent_kwargs,
+        memory=memory
+    )   
+    
+    return agent
