+++ config.toml
@@ -1,2 +1,2 @@
 [server]
-pythonExecutable = "research_agent_env\Scripts\pygmentize.exe"
+pythonExecutable = "research_agent_env\Scripts\python.exe"
